+++ Si_UnitBalance/Si_UnitBalance.Discord.cs
@@ -223,7 +223,6 @@ namespace Si_UnitBalance
             // Bool toggles
             CompareBoolSetting("additional_spawn", "Additional Spawn", defaultCfg, activeCfg, lastPushCfg, changes);
             CompareBoolSetting("shrimp_disable_aim", "Shrimp Disable Aim", defaultCfg, activeCfg, lastPushCfg, changes);
-            CompareBoolSetting("revert_on_round_end", "Revert on Round End", defaultCfg, activeCfg, lastPushCfg, changes);
             CompareBoolSetting("health_mult_enabled", "Health Mult (server-only)", defaultCfg, activeCfg, lastPushCfg, changes);
         }